Add some more tests and make it working.
